// src/enums/vehicleType.enum.ts
export enum VehicleTypeEnum {
	SUV = "SUV",
	PICKUP = "PICKUP",
	MINIVAN = "MINIVAN",
	LIGHT_TRUCK = "LIGHT_TRUCK",
	COUPE = "COUPE",
	HATCHBACK = "HATCHBACK",
	FURGON = "FURGON",
	SEDAN = "SEDAN",
	VAN = "VAN",
	RURAL = "RURAL",
	CABRIOLET = "CABRIOLET",
	SPORTSCAR = "SPORTSCAR",
	ROADSTER = "ROADSTER",
	ELECTRIC = "ELECTRIC",
	HYBRID = "HYBRID",
	LUXURY = "LUXURY",
	OFF_ROAD = "OFF_ROAD",
	PICKUP_TRUCK = "PICKUP_TRUCK",
	CROSSOVER = "CROSSOVER",
	COMPACT = "COMPACT",
}